requirements.txt
fastapi
uvicorn
openai==0.27.8
streamlit
requests
python-dotenv
six